OSIS-4542 : Refactor
