+++ education_group/api/serializers/group.py
@@ -34,37 +34,20 @@ from base.models.education_group_year import EducationGroupYear
 from base.models.enums import education_group_categories
 from education_group.api.serializers import utils
 from education_group.api.serializers.education_group_title import EducationGroupTitleSerializer
+from education_group.api.serializers.education_group_year import BaseEducationGroupYearSerializer
 from education_group.api.serializers.utils import GroupHyperlinkedIdentityField
 
 
-class GroupDetailSerializer(EducationGroupTitleSerializer, serializers.ModelSerializer):
+class GroupDetailSerializer(BaseEducationGroupYearSerializer, serializers.ModelSerializer):
     url = GroupHyperlinkedIdentityField(read_only=True)
-    code = serializers.CharField(source='partial_acronym', read_only=True)
-    academic_year = serializers.SlugRelatedField(slug_field='year', queryset=AcademicYear.objects.all())
-    education_group_type = serializers.SlugRelatedField(
-        slug_field='name',
-        queryset=EducationGroupType.objects.filter(category=education_group_categories.GROUP),
-    )
-    management_entity = serializers.CharField(source='management_entity_version.acronym', read_only=True)
-    management_faculty = serializers.SerializerMethodField()
     remark = serializers.SerializerMethodField()
     campus = CampusDetailSerializer(source='main_teaching_campus', read_only=True)
 
     # Display human readable value
-    education_group_type_text = serializers.CharField(source='education_group_type.get_name_display', read_only=True)
     constraint_type_text = serializers.CharField(source='get_constraint_type_display', read_only=True)
 
-    class Meta(EducationGroupTitleSerializer.Meta):
-        model = EducationGroupYear
-        fields = EducationGroupTitleSerializer.Meta.fields + (
-            'url',
-            'acronym',
-            'code',
-            'management_entity',
-            'management_faculty',
-            'academic_year',
-            'education_group_type',
-            'education_group_type_text',
+    class Meta(BaseEducationGroupYearSerializer.Meta):
+        fields = ('url', ) + BaseEducationGroupYearSerializer.Meta.fields + (
             'credits',
             'min_constraint',
             'max_constraint',
@@ -80,7 +63,3 @@ class GroupDetailSerializer(EducationGroupTitleSerializer, serializers.ModelSeri
             education_group_year,
             'remark' + ('_english' if language and language not in settings.LANGUAGE_CODE_FR else '')
         )
-
-    @staticmethod
-    def get_management_faculty(obj):
-        return utils.get_entity(obj, 'management')
